Move functionality of betweenness statistics to dybc main library
