+++ src/exp/dybc_batch_update.c
@@ -4,7 +4,7 @@
 #include <assert.h>
 #include <igraph/igraph.h>
 
-#include "dybc_update_query.h"
+#include "dybc/dybc_update_query.h"
 #include "dybc_exp_common.h"
 
 void batch_update(igraph_t* G,
