+++ src/exp/dybc_batch_update.h
@@ -3,7 +3,7 @@
 #define _DYBC_BATCH_UPDATE_H_
 
 #include <igraph/igraph.h>
-#include "dybc_update_query.h"
+#include "dybc/dybc_update_query.h"
 
 void batch_update(igraph_t* G,
                   igraph_matrix_t* D,
