+++ src/exp/exp-betweenness-stats.c
@@ -7,8 +7,8 @@
 #include <igraph/igraph_math.h>
 #include "dybc_exp_common.h"
 #include "dybc_io.h"
-#include "dybc_update_query.h"
-#include "dybc_update_stats.h"
+#include "dybc/dybc_update_query.h"
+#include "dybc/dybc_update_stats.h"
 
 const char *argp_program_version =
   "exp-betweenness-stats beta";
